Make configuration client-specific
This allows the main service client and the operations client
to have separate configurations.
