--- gapic-generator/templates/default/service/config.erb
@@ -1,6 +1,0 @@
-<%- assert_locals service -%>
-<%= render partial: "service/client/config",
-           layout: "layouts/ruby",
-           locals: { service: service,
-                     namespaces: service.namespaces }
-%>
