+++ shared/output/cloud/showcase/lib/google/showcase/v1alpha3/testing/client.rb
@@ -18,7 +18,6 @@ require "google/gax"
 
 require "google/showcase/version"
 require "google/showcase/v1alpha3/testing_pb"
-require "google/showcase/v1alpha3/testing/configure"
 require "google/showcase/v1alpha3/testing/credentials"
 require "google/showcase/v1alpha3/testing/paths"
 
@@ -32,6 +31,72 @@ module Google
 
           # @private
           attr_reader :testing_stub
+
+          ##
+          # Configuration for the Testing API.
+          #
+          def self.configure
+            @configure ||= Google::Gax::Configuration.new do |config|
+              default_scope = Google::Gax::Configuration.deferred do
+                Credentials::SCOPE
+              end
+              config.add_field! :host,         "localhost", match: [String]
+              config.add_field! :port,         7469, match: [Integer]
+              config.add_field! :scope,        default_scope,                         match: [String, Array], allow_nil: true
+              config.add_field! :lib_name,     nil,                                   match: [String],        allow_nil: true
+              config.add_field! :lib_version,  nil,                                   match: [String],        allow_nil: true
+              config.add_field! :interceptors, [],                                    match: [Array]
+
+              config.add_field! :timeout,     60,  match: [Numeric]
+              config.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+              config.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+
+              config.add_config! :methods do |methods|
+                methods.add_config! :create_session do |method|
+                  method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                  method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                  method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                end
+                methods.add_config! :get_session do |method|
+                  method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                  method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                  method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                end
+                methods.add_config! :list_sessions do |method|
+                  method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                  method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                  method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                end
+                methods.add_config! :delete_session do |method|
+                  method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                  method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                  method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                end
+                methods.add_config! :report_session do |method|
+                  method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                  method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                  method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                end
+                methods.add_config! :list_tests do |method|
+                  method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                  method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                  method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                end
+                methods.add_config! :delete_test do |method|
+                  method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                  method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                  method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                end
+                methods.add_config! :verify_test do |method|
+                  method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                  method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                  method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                end
+              end
+            end
+            yield @configure if block_given?
+            @configure
+          end
 
           ##
           # Configure the Client client.
@@ -78,7 +143,7 @@ module Google
             require "google/showcase/v1alpha3/testing_services_pb"
 
             # Create the configuration object
-            config ||= Testing.configure
+            config ||= Client.configure
             config = config.derive! unless config.derived?
 
             # Yield the configuration if needed
