+++ shared/output/cloud/speech/lib/google/cloud/speech/v1/speech/client.rb
@@ -18,7 +18,6 @@ require "google/gax"
 
 require "google/cloud/speech/version"
 require "google/cloud/speech/v1/cloud_speech_pb"
-require "google/cloud/speech/v1/speech/configure"
 require "google/cloud/speech/v1/speech/credentials"
 require "google/cloud/speech/v1/speech/operations"
 
@@ -31,6 +30,47 @@ module Google
           class Client
             # @private
             attr_reader :speech_stub
+
+            ##
+            # Configuration for the Speech API.
+            #
+            def self.configure
+              @configure ||= Google::Gax::Configuration.new do |config|
+                default_scope = Google::Gax::Configuration.deferred do
+                  Credentials::SCOPE
+                end
+                config.add_field! :host,         "speech.googleapis.com", match: [String]
+                config.add_field! :port,         443, match: [Integer]
+                config.add_field! :scope,        default_scope,                         match: [String, Array], allow_nil: true
+                config.add_field! :lib_name,     nil,                                   match: [String],        allow_nil: true
+                config.add_field! :lib_version,  nil,                                   match: [String],        allow_nil: true
+                config.add_field! :interceptors, [],                                    match: [Array]
+
+                config.add_field! :timeout,     60,  match: [Numeric]
+                config.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                config.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+
+                config.add_config! :methods do |methods|
+                  methods.add_config! :recognize do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                  methods.add_config! :long_running_recognize do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                  methods.add_config! :streaming_recognize do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                end
+              end
+              yield @configure if block_given?
+              @configure
+            end
 
             ##
             # Configure the Client client.
@@ -77,7 +117,7 @@ module Google
               require "google/cloud/speech/v1/cloud_speech_services_pb"
 
               # Create the configuration object
-              config ||= Speech.configure
+              config ||= Client.configure
               config = config.derive! unless config.derived?
 
               # Yield the configuration if needed
@@ -103,8 +143,7 @@ module Google
               end
 
               @operations_client = Operations.new(
-                credentials: credentials,
-                config:      @config
+                credentials: credentials
               )
 
               @speech_stub = Google::Gax::Grpc::Stub.new(
