+++ shared/output/gapic/templates/speech/lib/google/cloud/speech/v1/speech/operations.rb
@@ -40,6 +40,60 @@ module Google
             attr_reader :operations_stub
 
             ##
+            # Configuration for the Operations API.
+            #
+            def self.configure
+              @configure ||= Google::Gax::Configuration.new do |config|
+                default_scope = Google::Gax::Configuration.deferred do
+                  Credentials::SCOPE
+                end
+                config.add_field! :host,         "speech.googleapis.com", match: [String]
+                config.add_field! :port,         443, match: [Integer]
+                config.add_field! :scope,        default_scope,                         match: [String, Array], allow_nil: true
+                config.add_field! :lib_name,     nil,                                   match: [String],        allow_nil: true
+                config.add_field! :lib_version,  nil,                                   match: [String],        allow_nil: true
+                config.add_field! :interceptors, [],                                    match: [Array]
+
+                config.add_field! :timeout,     60,  match: [Numeric]
+                config.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                config.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+
+                config.add_config! :methods do |methods|
+                  methods.add_config! :list_operations do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                  methods.add_config! :get_operation do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                  methods.add_config! :delete_operation do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                  methods.add_config! :cancel_operation do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                end
+              end
+              yield @configure if block_given?
+              @configure
+            end
+
+            ##
+            # Configure the Client client.
+            #
+            def configure
+              yield @config if block_given?
+              @config
+            end
+
+            ##
             # @param credentials [Google::Auth::Credentials, String, Hash,
             #   GRPC::Core::Channel, GRPC::Core::ChannelCredentials, Proc]
             #   Provides the means for authenticating requests made by the client. This
@@ -76,7 +130,7 @@ module Google
               require "google/longrunning/operations_services_pb"
 
               # Create the configuration object
-              config ||= Speech.configure
+              config ||= Operations.configure
               config = config.derive! unless config.derived?
 
               # Yield the configuration if needed
