+++ shared/output/gapic/templates/vision/lib/google/cloud/vision/v1/product_search/client.rb
@@ -26,7 +26,6 @@ require "google/gax"
 
 require "google/cloud/vision/version"
 require "google/cloud/vision/v1/product_search_service_pb"
-require "google/cloud/vision/v1/product_search/configure"
 require "google/cloud/vision/v1/product_search/credentials"
 require "google/cloud/vision/v1/product_search/operations"
 
@@ -39,6 +38,122 @@ module Google
           class Client
             # @private
             attr_reader :product_search_stub
+
+            ##
+            # Configuration for the ProductSearch API.
+            #
+            def self.configure
+              @configure ||= Google::Gax::Configuration.new do |config|
+                default_scope = Google::Gax::Configuration.deferred do
+                  Credentials::SCOPE
+                end
+                config.add_field! :host,         "vision.googleapis.com", match: [String]
+                config.add_field! :port,         443, match: [Integer]
+                config.add_field! :scope,        default_scope,                         match: [String, Array], allow_nil: true
+                config.add_field! :lib_name,     nil,                                   match: [String],        allow_nil: true
+                config.add_field! :lib_version,  nil,                                   match: [String],        allow_nil: true
+                config.add_field! :interceptors, [],                                    match: [Array]
+
+                config.add_field! :timeout,     60,  match: [Numeric]
+                config.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                config.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+
+                config.add_config! :methods do |methods|
+                  methods.add_config! :create_product_set do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                  methods.add_config! :list_product_sets do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                  methods.add_config! :get_product_set do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                  methods.add_config! :update_product_set do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                  methods.add_config! :delete_product_set do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                  methods.add_config! :create_product do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                  methods.add_config! :list_products do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                  methods.add_config! :get_product do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                  methods.add_config! :update_product do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                  methods.add_config! :delete_product do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                  methods.add_config! :create_reference_image do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                  methods.add_config! :delete_reference_image do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                  methods.add_config! :list_reference_images do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                  methods.add_config! :get_reference_image do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                  methods.add_config! :add_product_to_product_set do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                  methods.add_config! :remove_product_from_product_set do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                  methods.add_config! :list_products_in_product_set do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                  methods.add_config! :import_product_sets do |method|
+                    method.add_field! :timeout,     nil, match: [Numeric],    allow_nil: true
+                    method.add_field! :metadata,    nil, match: [Hash],       allow_nil: true
+                    method.add_field! :retry_codes, nil, match: [Hash, Proc], allow_nil: true
+                  end
+                end
+              end
+              yield @configure if block_given?
+              @configure
+            end
 
             ##
             # Configure the Client client.
@@ -85,7 +200,7 @@ module Google
               require "google/cloud/vision/v1/product_search_service_services_pb"
 
               # Create the configuration object
-              config ||= ProductSearch.configure
+              config ||= Client.configure
               config = config.derive! unless config.derived?
 
               # Yield the configuration if needed
@@ -111,8 +226,7 @@ module Google
               end
 
               @operations_client = Operations.new(
-                credentials: credentials,
-                config:      @config
+                credentials: credentials
               )
 
               @product_search_stub = Google::Gax::Grpc::Stub.new(
